moving the problem class into its own file
