+++ optimism/__init__.py
@@ -15,7 +15,8 @@ config.update("jax_platforms", "cpu")
 del config
 
 from .material import *
-from .Domain import Domain, Problem
+from .Domain import Domain
 from .EquationSolver import get_settings
 from .FunctionSpace import EssentialBC
+from .Problem import Problem
 from .ReadExodusMesh import read_exodus_mesh
